setting up email module
