+++ routes/gigs-route.js
@@ -12,6 +12,7 @@ let zlib = require('zlib');
 let s3 = new AWS.S3();
 
 var nodeMailer = require('nodemailer');
+var mailer = require(__dirname + '/../customModules/email');
 
 module.exports = (apiRouter) => {
   apiRouter.route('/gigs')
@@ -32,22 +33,27 @@ module.exports = (apiRouter) => {
           User.findByIdAndUpdate(userInfo, { $push: {gigs: gig._id}}, (err, user) => {
           })
 
-          var transporter = nodeMailer.createTransport('smtps://snapgignotification@gmail.com:snapsnap@smtp.gmail.com');
-          console.log('SENDING EMAIL :');
-          var mailOptions = {
-            from: '"snapgig" <snapgignotification@gmail.com>',
-            to: req.user.email,
-            subject: 'New Gig, ' + req.body.name + ' has been created',
-            text: req.user.username + ', your gig has been successfully submitted. The deadline for submissions is, ' + req.body.deadline
-          }
-          transporter.sendMail(mailOptions, function (error, info) {
-            if (error) {
-              res.json({msg: 'Gig couldn\'t be posted'});
-              res.end();
-            }
-            console.log('EMAIL HAS BEEN SEND', info);
-            res.status(200).json({data: gig});
-            res.end()
+          // var transporter = nodeMailer.createTransport('smtps://snapgignotification@gmail.com:snapsnap@smtp.gmail.com');
+          // console.log('SENDING EMAIL :');
+          // var mailOptions = {
+          //   from: '"snapgig" <snapgignotification@gmail.com>',
+          //   to: req.user.email,
+          //   subject: 'New Gig, ' + req.body.name + ' has been created',
+          //   text: req.user.username + ', your gig has been successfully submitted. The deadline for submissions is, ' + req.body.deadline
+          // }
+          // transporter.sendMail(mailOptions, function (error, info) {
+          //   if (error) {
+          //     res.json({msg: 'Gig couldn\'t be posted'});
+          //     res.end();
+          //   }
+          //   console.log('EMAIL HAS BEEN SEND', info);
+          //   res.status(200).json({data: gig});
+          //   res.end()
+          // })
+          mailer.gig(req.user.email, req.body.name, req.user.username, req.body.deadline, gig, (err, info) => {
+            if (err) throw err;
+            res.json({data: gig})
+            res.end();
           })
         })
       })
@@ -93,6 +99,7 @@ module.exports = (apiRouter) => {
   apiRouter.route('/gigs/:id/submissions')
     .post((req, res) => {
       req.on('data', (data) => {
+        console.log('REQUEST USER AFTER HITTING ROUTE : ', req.user);
         var newBody;
         req.body = JSON.parse(data);
         let newSub = new Sub(req.body);
